fix(ci): pin smtp4dev (#32035)
* fix(ci): pin smtp4dev

New version seems to break something.
