+++ frappe/public/js/frappe/doctype/index.js
@@ -222,51 +222,4 @@ frappe.model.DocTypeController = class DocTypeController extends frappe.ui.form.
 			update_fieldname_options();
 		}
 	}
-
-	async trim_table() {
-		let frm = this.frm;
-		let doctype = null;
-		if (frm.doc.doctype === "DocType") {
-			doctype = frm.doc.name;
-		} else {
-			// In customize form name field is "Customize Form". Doctype name is stored in doc_type.
-			doctype = frm.doc.doc_type;
-		}
-
-		let dropped_columns = await frappe.xcall(
-			"frappe.custom.doctype.customize_form.customize_form.get_orphaned_columns",
-			{ doctype: doctype }
-		);
-
-		if (!dropped_columns?.length) {
-			frappe.toast(__("This doctype has no orphan fields to trim"));
-			return;
-		}
-		let msg = __(
-			"Warning: DATA LOSS IMMINENT! Proceeding will permanently delete following database columns from doctype {0}:",
-			[frm.doc.name.bold()]
-		);
-		msg += "<ol>" + dropped_columns.map((col) => `<li>${col}</li>`).join("") + "</ol>";
-		msg += __("This action is irreversible. Do you wish to continue?");
-
-		frappe.confirm(msg, () => {
-			return frm.call({
-				doc: frm.doc,
-				method: "trim_table",
-				callback: function (r) {
-					if (!r.exc) {
-						frappe.show_alert({
-							message: __("Table Trimmed"),
-							indicator: "green",
-						});
-						if (frm.doc.doctype === "DocType") {
-							frm.refresh();
-						} else {
-							frappe.customize_form.clear_locals_and_refresh(frm);
-						}
-					}
-				},
-			});
-		});
-	}
 };
